Move code to sql.py and field_ext.py
