+++ tortoise_pathway/operations/__init__.py
@@ -14,7 +14,7 @@ from tortoise_pathway.operations.drop_index import DropIndex
 from tortoise_pathway.operations.drop_model import DropModel
 from tortoise_pathway.operations.rename_field import RenameField
 from tortoise_pathway.operations.rename_model import RenameModel
-from tortoise_pathway.operations.utils import field_to_migration
+from tortoise_pathway.operations.field_ext import field_to_migration
 
 __all__ = [
     "Operation",
